리뷰보드 재수정

+++ src/main/java/com/itwill/dao/ReviewBoardDao.java
@@ -20,7 +20,7 @@ public interface ReviewBoardDao {
 
 	List<ReviewBoard> findByProductNo(Long productNo); // productNo로 reviewboard 리스트 검색
 
-	public List<ReviewBoard> findByStarAll(Long star); //선택한 별점으로 찾기
+	public List<ReviewBoard> findByStarAll(Double star); //선택한 별점으로 찾기
 
 	 public List<ReviewBoard> findByUserNo(Long no); // 선택된 userId 리뷰 리스트만 나오기
